requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain-openai==0.1.1
google-cloud-firestore==2.13.1
stripe==7.6.0
python-dotenv==1.0.0